알람 이벤트까지 개발

+++ src/main/java/com/fastcampus/sns/controller/PostController.java
@@ -71,7 +71,7 @@ public class PostController {
     }
 
     @GetMapping("/{postId}/likes")
-    public Response<Integer> likeCount(
+    public Response<Long> likeCount(
             @PathVariable("postId") Integer postId,
             Authentication authentication
     ) {
